Fold payments formatters into the feature helpers module
Move the date/currency formatters out of the standalone format.ts and into the
existing #features/payments/helpers module (alongside the environment-value
helpers), repointing the four payment pages and renaming the test to
helpers.test.ts. No behavior change.
